add devshell-files-cmd which can be used outside of the context of devshell
It differs from the existing devshell-files command with by the use of
full nix paths of the individual file scripts as opposed to referring
to them by via $DEVSHELL_DIR. This is more portable as it doesn't depend
on the context being setup outside of the script.

NOTE: I've removed the . from the prefix of the file script names (.dsf => dsf) so
lib.getExe works on the script the derivations. This could be reverted is there
is a good reason for the . to be there.